Some additional thoughts to OSGi post 2019
This document - I think - is a crucial piece for the future of OSGi. 
While there is a clear case for the tech in IoT (most likely because thats where OSGi originates from) it has lost ground in the recent years in areas that are now occupied by Spring(Boot), Micronaut,Microprofile,NodeJS. 
As a consultant, I have barely seen new projects outside the embedded space who successfully adopted OSGi in its full glory. 
And the reasons are not so technical as described in this document (although i'd agree with almost everything said here). 
Let me open some additional thoughts, we need to see how to put this in structure:
* Pace of change is very slow because of the nature of specifications. Compared to Springs pace, OSGi DS is creeping slow. 
* Module Layer is awesome when you need the reloading aspect, but few actually use it. (see immutable containers).
* Module Layer helpers solving conflicting transitive dependency problems (e.g. our friend guava that every 2nd lib uses in a different version). But it comes at a huge cost: its annoying to non-experts trying to integrate legacy libraries (libs doing classloader magic themselves).
* I recently came to the point to refer to OSGi as "the school of modular software design". What I mean by that is that OSGi is so hard to get good at that you are a better developer (better person??) afterwards. We might need to rethink if this is how the tech is perceived. It has its value, I just don't see OSGi/the alliance drive towards this goal. 
* As written in the document, not everyone has Netflix scale, actually most don't. I'd like to see OSGi find its spot (next to IoT) and optimise for that. I'd be very interested in optimising a subset (e.g. service layer only, or even just directly DS) onto substrateVM. An adaptive service registry is something that other frameworks don't offer yet. But part of me thinks that this might be done best as an OSGi fork - see all the other issues. On the other hand - maybe there is still enough drive to do this within OSGi official?